+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/1442>)
 - Get target information for tabular dataset
   (<https://github.com/openvinotoolkit/datumaro/pull/1471>)
+- Add ExtractedMask and update importers who can use it to use it
+  (<https://github.com/openvinotoolkit/datumaro/pull/1480>)
 
 ### Bug fixes